requirements.txt
pymongo
PyMuPDF==1.22.5
psycopg2-binary
ollama 
tqdm
